+++ docs/packages/interactive/element/README.md.c
@@ -7,6 +7,7 @@
 
 - 进入页面时获取节点对象实例，在 Page.onReady 中使用
 - 通过 setData 切换组件状态后获取节点实例，在 setData 回调中使用
+- 在 Rax 运行时场景下，可通过`window.addEventListener('setDataFinished', () => {});`的回调中调用，确保节点可以获取到
 
 ## 支持
 <img alt="browser" src="https://gw.alicdn.com/tfs/TB1uYFobGSs3KVjSZPiXXcsiVXa-200-200.svg" width="25px" height="25px" /> <img alt="miniApp" src="https://gw.alicdn.com/tfs/TB1bBpmbRCw3KVjSZFuXXcAOpXa-200-200.svg" width="25px" height="25px" title="ali miniprogram" /> <img alt="wechatMiniprogram" src="https://img.alicdn.com/tfs/TB1slcYdxv1gK0jSZFFXXb0sXXa-200-200.svg" width="25px" height="25px" title="wechatMiniprogram" /> <img alt="bytedanceMicroApp" src="https://gw.alicdn.com/tfs/TB1jFtVzO_1gK0jSZFqXXcpaXXa-200-200.svg" width="25px" height="25px" title="bytedanceMicroApp" /> <img alt="baiduSmartProgram" src="https://img.alicdn.com/imgextra/i4/O1CN01jngdBb24yGv2Fu34G_!!6000000007459-2-tps-200-200.png" width="25px" height="25px" title="百度小程序" /> <img alt="kuaiShouMiniProgram" src="https://gw.alicdn.com/imgextra/i4/O1CN01kzmJMM24jcFEzp5Wv_!!6000000007427-2-tps-200-200.png" width="25px" height="25px" title="快手小程序" />
@@ -25,14 +26,16 @@ import { getScrollOffset, getBoundingClientRect } from '@uni/element';
 
 function App() {
   useEffect(() => {
-    getScrollOffset('#container').then((ret) => {
-      const { scrollTop, scrollLeft } = ret[0];
-      console.log(scrollTop, scrollLeft);
-    });
+    window.addEventListener('setDataFinished', () => {
+      getScrollOffset('#container').then((ret) => {
+        const { scrollTop, scrollLeft } = ret[0];
+        console.log(scrollTop, scrollLeft);
+      });
 
-    getBoundingClientRect('#container').then((ret) => {
-      const { width, height, top, left, right, bottom } = ret[0];
-      console.log(width, height, top, left, right, bottom);
+      getBoundingClientRect('#container').then((ret) => {
+        const { width, height, top, left, right, bottom } = ret[0];
+        console.log(width, height, top, left, right, bottom);
+      });
     });
   }, []);