+++ src/packages/device/vibrate/src/common.ts
@@ -1,0 +1,9 @@
+/* eslint-disable eqeqeq */
+import { promisify } from '@utils/promisify';
+import { styleIn } from '@utils/styleOptions';
+
+export function normalize(api, containerName) {
+  return (args?) => {
+    return promisify(api)(styleIn(args, containerName));
+  };
+}