+++ src/packages/device/vibrate/src/types.ts
@@ -1,0 +1,9 @@
+export interface Params {
+  success?: (res?) => any;
+  fail?: (res?) => any;
+  complete?: (res?) => any;
+}
+
+export interface ShortPARAMS extends Params {
+  type?: 'heavy' | 'medium' | 'light';
+}